% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn.R
\name{attachLabels}
\alias{attachLabels}
\title{Attach labels to vennobjects}
\usage{
attachLabels(vo, cat = c(vo$A$label, vo$B$label), cat.labelpos = NULL,
  cat.angle = NULL, cat.pos = c(vo$A$pos, vo$B$pos),
  cat.offset = c(vo$A$offset, vo$B$offset))
}
\arguments{
\item{vo}{- vennobject.}

\item{cat}{- character vector. Two labels for circles A and B}

\item{cat.labelpos}{-}

\item{cat.angle}{- if absolute position is not specified,
can specify angle (in radians) from center
0 corresponds to label just above the circle.}

\item{cat.pos}{- a vector with four items (adjA, adjA, adjB, adjB) use this with cat.angle to
tune where label appears relative to the  anchor position}

\item{cat.offset}{- numeric vector, distance of labels from}
}
\description{
This function saves category labels into a venn object.
If the object is moved, the labels will also move.
}

