% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colLighten.R
\name{avgcol}
\alias{avgcol}
\title{Calculate an average from a vector of colors}
\usage{
avgcol(x)
}
\arguments{
\item{x}{vector of colors.}
}
\description{
Calculate an average from a vector of colors.
}

