% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveLatexTable.R
\name{avoidLatexSpecialChar}
\alias{avoidLatexSpecialChar}
\title{takes an object, vector, or matrix as input
if the object is made up characters, latex special characters are edited}
\usage{
avoidLatexSpecialChar(Data)
}
\arguments{
\item{Data}{data frame}
}
\description{
takes an object, vector, or matrix as input
if the object is made up characters, latex special characters are edited
}

