% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catlog.R
\name{catlog}
\alias{catlog}
\title{Print a log message}
\usage{
catlog(...)
}
\arguments{
\item{...}{any number of elements to print beside the timestamp}
}
\description{
Print a message including a timestamp. The output will include a newline
character at the end.
}

