% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catlog.R
\name{catlogreturn}
\alias{catlogreturn}
\title{Print and return a log message}
\usage{
catlogreturn(...)
}
\arguments{
\item{...}{any number of elements to print beside the timestamp}
}
\description{
Print a message including a timestamp. Also return the content of the
message. The output includes a newline character at the end.
}

