% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkVarsStop.R
\name{checkVarsStop}
\alias{checkVarsStop}
\title{Compare vectors and report discrepancies}
\usage{
checkVarsStop(have, need)
}
\arguments{
\item{have}{vector of elements. Call the function passing ls() here.}

\item{need}{vector of elements. Call the function specifying all
the required variables here.}
}
\description{
Compare vectors and report discrepancies via stop(). This is useful
at the beginning of a pipeline, where it is important to check whether
critical variables/inputs have been defined.
}

