% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colLighten.R
\name{col2hex}
\alias{col2hex}
\title{Convert colors into hex format}
\usage{
col2hex(x)
}
\arguments{
\item{x}{color or vector of colors to convert}
}
\description{
Convert colors into hex format, e.g. "red" into #ff0000.
}

