% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colLighten.R
\name{colDarken}
\alias{colDarken}
\title{Modify a color by darkening}
\usage{
colDarken(col, darken)
}
\arguments{
\item{col}{color to modify. The funciton also accepts vectors of colors.}

\item{dark}{number [0, whatever] indicating by how much to lighten.}
}
\description{
Modify a color by darkening. Analogous to colLighten() in this package.
}

