% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colLighten.R
\name{colLighten}
\alias{colLighten}
\title{Modify a color by lightening}
\usage{
colLighten(col, lighten)
}
\arguments{
\item{col}{color to modify. The funciton also accepts vectors of colors.}

\item{lighten}{number [0, whatever] indicating by how much to lighten.}
}
\description{
Modify a color by lightening. Analogous to colDarken() in this package.
}

