% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compileLatexDoc.R
\name{compileLatexDoc}
\alias{compileLatexDoc}
\title{Compile a latex document and}
\usage{
compileLatexDoc(filename, latexmode = "latex", intern = TRUE)
}
\arguments{
\item{filename}{character string, name of latex document}

\item{latexmode}{character string,  should be either "latex" or "pdflatex"
Both modes will give  a pdf document at the end,
but use "latex" if document has eps figures (tex -> dvi -> pdf)
and use "pdflatex" if document has pdf figures (tex -> pdf)}

\item{intern}{logical, passed to a system() command}
}
\description{
Compile a latex document and
}

