% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templateFiller.R
\name{fill.templates}
\alias{fill.templates}
\title{Fill-in a template with specific key-value pairs}
\usage{
fill.templates(templs, templatename, keys = c(), values = c())
}
\arguments{
\item{templs}{list of templates}

\item{templatename}{character string, name of template to use}

\item{keys}{vector of character strings, elements in the template to replace}

\item{values}{vector of character strings, matched with keys, these values
 will be used to replace generic keys within the template}
}
\description{
Fill-in a template with specific key-value pairs
}

