% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn.R
\name{getVenn}
\alias{getVenn}
\title{Create a vennobject}
\usage{
getVenn(sizeA, sizeB, sizeAB, theta = 0, ...)
}
\arguments{
\item{sizeA}{numeric. Area of circle A}

\item{sizeB}{numeric. Area of circle B}

\item{sizeAB}{numeric. Area of intersection}

\item{theta}{numeric. Angle in radians}

\item{...}{additional arguments passed on to findCircleDistance}
}
\description{
Convert areas into coordinates of circle centers
}

