% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn.R
\name{getVennLims.vennobject}
\alias{getVennLims.vennobject}
\title{Obtain xlim and ylim for vennobject}
\usage{
getVennLims.vennobject(vo)
}
\arguments{
\item{vo}{- vennobject}
}
\description{
Obtain xlim and ylim for vennobject
}

