% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadOneObject.R
\name{loadOneObject}
\alias{loadOneObject}
\title{Load one R object from and Rda file}
\usage{
loadOneObject(file)
}
\arguments{
\item{file}{filename with Rda object to load}
}
\description{
Function loads an object from an Rda file. In contrast to standard
load(), this function returns the loaded object as opposed to put it
into the namespace. Thus, one can place the new object into any variable,
not necessarily the variable name used to save it in the first place.
}

