% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeTempDir.R
\name{makeTempDir}
\alias{makeTempDir}
\title{Create a temporary directory}
\usage{
makeTempDir(basedir, label, id = NULL)
}
\arguments{
\item{basedir}{- location where the temporary directory will be placed}

\item{label}{- a prefix for the temporary directory name}

\item{id}{- a suffix for the temporary directory. If left NULL, a random number is used}
}
\description{
Creates a tempoary direcoty. Returns path to this new directory
}

