% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn.R
\name{moveVennObject}
\alias{moveVennObject}
\title{displace the Venn diagram by a vector}
\usage{
moveVennObject(vo, moveby)
}
\arguments{
\item{vo}{- vennobject}

\item{moveby}{- numeric vector of length 2.}
}
\description{
displace the Venn diagram by a vector
}

