% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPS.R
\name{plotPDFend}
\alias{plotPDFend}
\title{Terminate writing into a PDF device.}
\usage{
plotPDFend(file = NULL, ...)
}
\arguments{
\item{file}{filename. When NULL is not specified (left NULL),
the function completes without doing anything. When filename is
not NULL, this function will force a call to dev.off() and complete
writing to a graphics device.}
}
\description{
Terminate writing into a PDF device.
}

