% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPS.R
\name{plotPSstart}
\alias{plotPSstart}
\title{Start plotting into a PS device}
\usage{
plotPSstart(file = NULL, fsize = c(4, 4), ...)
}
\arguments{
\item{file}{filename where to save the image. When this is not
specified (left NULL), the function completes without doing anything.}

\item{fsize}{vector with two elements giving (width, height)
of desired image}
}
\description{
Start plotting into a PS device.
}

