% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn.R
\name{print.vennobject}
\alias{print.vennobject}
\title{Show basic information about a vennobject}
\usage{
\method{print}{vennobject}(x, ...)
}
\arguments{
\item{x}{vennobject}

\item{...}{Further parameters are ignored}
}
\description{
Show basic information about a vennobject
}

