% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PSZ.R
\name{PSZ}
\alias{PSZ}
\title{A wrapper for paste()}
\usage{
PSZ(...)
}
\arguments{
\item{...}{any number of objects to be concatenated into one string}
}
\description{
Function concatenates string without putting in additional spaces.
Name of function stands for Pasta Sep Zero.
}

