% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templateFiller.R
\name{read.templates}
\alias{read.templates}
\title{Read template objects from files}
\usage{
read.templates(dirname)
}
\arguments{
\item{dirname}{character string, name of directory holding templates}
}
\description{
This function scans a directory for .txt files and loads each of them
into a list of strings
}

