% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveLatexTable.R
\name{saveLatexTable}
\alias{saveLatexTable}
\title{Save a matrix as latex table}
\usage{
saveLatexTable(Data, file = NULL, hlineevery = 5, checkSpecialChars = F,
  eol = " \\\\\\\\\\n", quote = FALSE, sep = " & ", col.names = F,
  row.names = F, ...)
}
\arguments{
\item{Data}{data frame}

\item{file}{character string, destination file name}

\item{hlineevery}{integer, number of lines between hline elements}

\item{checkSpecialChars}{logical, determine if special characters should be escaped}

\item{eol}{character string, text for end of line}

\item{quote}{logical,}

\item{sep}{character string, separator between columns}

\item{col.names}{logical,}

\item{row.names}{logical,}

\item{...}{additional arguments are ignored}
}
\description{
Save a data frame as a latex table. Convert object into a matrix
and write it into two files. One has suffix "header.txt" and contains
the column names. The second has suffix "body.txt" and contains the
contents of the table.
}

