% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showColors.R
\name{showColors}
\alias{showColors}
\title{Display a set of colors on screen}
\usage{
showColors(colors)
}
\arguments{
\item{colors}{vector of colors to display on screen}
}
\description{
Displays bars of various colors on screen. Each bar will be
annotated with the color name. Useful during development
for understanding how various colors will appear together.
}

