% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spearmanRho.R
\name{spearmanRho}
\alias{spearmanRho}
\title{Compute Spearman rho from two vectors}
\usage{
spearmanRho(x, y)
}
\arguments{
\item{x}{vector of numbers}

\item{y}{vector of numbers}
}
\description{
Compute Spearman rho from two vectors. Unlike cor.test(), this function
computes only the correlation estimate and does not bother with
the p-value. Skipping the p-value makes this function considerably
faster than cor.test().
}

