% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spearmanRho.R
\name{spearmanRhoFromRanks}
\alias{spearmanRhoFromRanks}
\title{Compute Spearman rho from ranks}
\usage{
spearmanRhoFromRanks(xranks, yranks)
}
\arguments{
\item{xranks}{vector of ranks}

\item{yranks}{vector of ranks.}
}
\description{
Compute Spearman rho from ranks. Use this function when comparing
two vectors for which the ranks are already known. E.g. avoid recomputing
the ranks over and over when performing multiple comparisons.
Unlike in cor.test(x, y), this function calculates only the rho value
and does not bother with the p-value.
}

