% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colLighten.R
\name{val2hex}
\alias{val2hex}
\title{Convert a value [0,1] into a hex code}
\usage{
val2hex(x)
}
\arguments{
\item{x}{a number or a vector of numbers to convert into hex.}
}
\description{
Convert a value [0,1] into a hex code. Useful for obtaining a transparency
code based on a floating point number.
}

