% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveLatexTable.R
\name{saveLatexTable}
\alias{saveLatexTable}
\title{Save a matrix as latex table}
\usage{
saveLatexTable(Data, file = NULL, hlineevery = 5, checkSpecialChars = F,
  eol = " \\\\\\\\\\n", quote = F, sep = " & ", col.names = F,
  row.names = F, ...)
}
\description{
Save a data frame as a latex table. Convert object into a matrix
and write it into two files. One has suffix "header.txt" and contains
the column names. The second has suffix "body.txt" and contains the
contents of the table.
}

